github_ready/requirements.txt
moviepy==1.0.3
numpy
Pillow
requests
